+++ infrastructure-common/src/main/kotlin/internal/ProjectProperties.kt
@@ -1,10 +1,9 @@
-package com.redmadrobot.build.kotlin.internal
+package com.redmadrobot.build.internal
 
-import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import org.gradle.api.Project
 
-@OptIn(InternalGradleInfrastructureApi::class)
-internal fun Project.findBooleanProperty(propertyName: String): Boolean? {
+@InternalGradleInfrastructureApi
+public fun Project.findBooleanProperty(propertyName: String): Boolean? {
     return findStringProperty(propertyName)?.toBoolean()
 }