+++ infrastructure-kotlin/src/main/kotlin/kotlin/internal/Kotlin.kt
@@ -4,6 +4,7 @@ import com.redmadrobot.build.InfrastructurePlugin
 import com.redmadrobot.build.dsl.isRunningOnCi
 import com.redmadrobot.build.dsl.kotlinCompile
 import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
+import com.redmadrobot.build.internal.findBooleanProperty
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.provider.Provider
@@ -21,6 +22,7 @@ public fun InfrastructurePlugin.configureKotlin(jvmTargetProperty: Provider<Java
 }
 
 /** Use property value if it exists or fallback to true if running on CI. */
+@OptIn(InternalGradleInfrastructureApi::class)
 private fun Project.getWarningsAsErrorsProperty(): Boolean {
     return findBooleanProperty("warningsAsErrors") ?: isRunningOnCi
 }